% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{add_acs_share}
\alias{add_acs_share}
\title{Add shares to Psrccensus ACS object}
\usage{
add_acs_share(df)
}
\arguments{
\item{df}{dataframe with Census ACS result}
}
\value{
dataframe with additional share and share_moe fields
}
\description{
Add shares to Psrccensus ACS object
}
