% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{add_county}
\alias{add_county}
\title{Add County Name to PUMS API Result}
\usage{
add_county(dt)
}
\arguments{
\item{dt}{PSRC data.table}
}
\value{
PSRC data.table with county names
}
\description{
Helper to \code{\link{psrc_pums_groupvar}} and \code{\link{psrc_pums_targetvar}} functions.
Attaches county name.
}
\author{
Michael Jensen
}
