% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{add_decennial_share}
\alias{add_decennial_share}
\title{Add shares to Psrccensus ACS object}
\usage{
add_decennial_share(df)
}
\arguments{
\item{df}{dataframe with Psrccensus Decennial result}
}
\value{
dataframe with additional share and share_moe fields
}
\description{
Add shares to Psrccensus ACS object
}
\author{
Michael Jensen
}
