% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{add_regional_acs}
\alias{add_regional_acs}
\title{Add regional totals}
\usage{
add_regional_acs(dt)
}
\arguments{
\item{dt}{data.table with Census API result}
}
\value{
dt with regional totals
}
\description{
Helper function for \code{\link{get_acs_single}} to add regional totals to county estimates
}
