% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{add_regional_decennial}
\alias{add_regional_decennial}
\title{Add regional totals}
\usage{
add_regional_decennial(dt)
}
\arguments{
\item{dt}{data.table with Census API result}
}
\value{
dt with regional totals
}
\description{
Helper function for \code{\link{get_decennial_single}} to add regional totals to county estimates
}
