% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{adjust_dollars}
\alias{adjust_dollars}
\title{Dollar variable adjustment for PUMS}
\usage{
adjust_dollars(dt)
}
\arguments{
\item{dt}{The data.table}
}
\value{
full input data.table with all dollar values adjusted
}
\description{
Helper to \code{\link{get_psrc_pums}} function.
Applies the Census Bureau-specified inflation adjustment to dollar values. See vignette for brief discussion.
Option to bypass this inflation adjustment exists, in order to match estimates generated without it.
}
\author{
Michael Jensen
}
