% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{adjust_inflation}
\alias{adjust_inflation}
\title{Inflation adjustment for PUMS}
\usage{
adjust_inflation(dt, dollar_var, adj_var)
}
\arguments{
\item{dt}{The data.table}

\item{dollar_var}{PUMS variable to adjust}

\item{adj_var}{PUMS adjustment factor: either ADJINC for income or ADJHSG for housing cost}
}
\value{
filtered input data.table with input variable adjusted
}
\description{
Helper to the \code{\link{adjust_dollars}} function
}
\author{
Michael Jensen
}
