% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{census_to_customgeo}
\alias{census_to_customgeo}
\title{Function to flexibly convert psrccensus estimates to any regional sf geography}
\usage{
census_to_customgeo(df, custom_geo, custom_geo_var, wgt = "total_pop")
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}

\item{custom_geo}{custom sf file with intended geography/geometry}

\item{custom_geo_var}{grouping variable from custom geography file, i.e. geography label}

\item{wgt}{measure share used as split weight
either "total_pop" (default), "household_pop", "group_quarters_pop", "housing_units" or "occupied_housing_units"}
}
\value{
table with custom geographic units in place of census geography units
}
\description{
Function to flexibly convert psrccensus estimates to any regional sf geography
}
\author{
Michael Jensen
}
