% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{clip2region}
\alias{clip2region}
\title{Restrict PUMS API pull to PSRC region}
\usage{
clip2region(dt)
}
\arguments{
\item{dt}{The data table}
}
\value{
the filtered data.table
}
\description{
This helper function is used in both the psrc_pums_groupvar and psrc_pums_targetvar functions.
It filters the Washington State API pull to the PSRC region and removes State/label columns
}
\author{
Michael Jensen
}
