% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{county_pums_stat}
\alias{county_pums_stat}
\alias{county_pums_total}
\alias{county_pums_count}
\alias{county_pums_median}
\alias{county_pums_mean}
\title{PUMS summary statistics by county}
\usage{
county_pums_total(so, target_var, group_vars = NULL)

county_pums_count(so, group_vars = NULL)

county_pums_median(so, target_var, group_vars = NULL)

county_pums_mean(so, target_var, group_vars = NULL)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{Factor variable/s for grouping, as an UPPERCASE string element or list}
}
\value{
A table with the counties, variable names and labels, summary statistic, and margin of error
}
\description{
Separate function for total, count, median, or mean
}
