% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus-helpers.R
\name{create_param_grid}
\alias{create_param_grid}
\title{Generate parameter list for all combinations}
\usage{
create_param_grid(items, years, params, item_name = "item")
}
\arguments{
\item{items}{Vector of table or data item names}

\item{years}{Vector of years}

\item{params}{Base parameters to include with each combination}

\item{item_name}{character string, i.e. "table" or "data_item"}
}
\value{
List of parameter combinations
}
\description{
Helper function for \code{\link{get_acs_recs}}; creates a list of parameter combinations for all tables and years
}
