% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tract-map.R
\name{create_tract_map}
\alias{create_tract_map}
\title{Census Tract Mapping}
\usage{
create_tract_map(
  tract.tbl,
  tract.lyr,
  tract.var,
  year,
  map.lat = 47.615,
  map.lon = -122.257,
  map.zoom = 8.5
)
}
\arguments{
\item{tract.tbl}{A data frame of census data by Census Tract to map.}

\item{tract.lyr}{A spatial layer of Census Tracts.}

\item{tract.var}{A character string or vector of attributes to map}

\item{year}{A numeric value equal or greater than 2010 to the latest ACS available year.}

\item{map.lat}{A numeric value for the latitude of the center point for your map. Defaults to 47.615 (PSRC Region)}

\item{map.lon}{A numeric value for the longitude of the center point for your map. Defaults to -122.257 (PSRC Region)}

\item{map.zoom}{A numeric value for the default zoom level for your map. Defaults to 8.5 (PSRC Region)}
}
\value{
a leaflet map object.
}
\description{
Generate Census Tract Maps for Census Data
}
\examples{
Sys.getenv("CENSUS_API_KEY")
\dontrun{tract.tbl <- psrccensus::get_acs_recs(geography='tract',table.names=c('B03002'),years=c(2019))
   tract.lyr <- sf::st_read(
   dsn="MSSQL:server=AWS-PROD-SQL\\\\Sockeye;database=ElmerGeo;trusted_connection=yes",
   layer="dbo.tract2010_nowater", crs=2285)
   create_tract_map<-function(tract.tbl=tract.tbl,
                           tract.lyr=tract.lyr,
                           tract.var=c("Estimate!!Total:!!Not Hispanic or Latino:!!Asian alone"),
                           year=2019)}

}
\author{
Suzanne Childress
}
