% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus-helpers.R
\name{decennial_varsearch}
\alias{decennial_varsearch}
\title{Search Decennial Census variables}
\usage{
decennial_varsearch(regex, year = NULL)
}
\arguments{
\item{regex}{search term}

\item{year}{optionally restrict search to a specific year}
}
\value{
data.table of filtered variable codes and attributes
}
\description{
Includes primary, subject and profile tables
}
\author{
Michael Jensen
}
