% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{fetch_ftp}
\alias{fetch_ftp}
\title{Fetch FTP}
\usage{
fetch_ftp(span, dyear, level)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{level}{Either "p" or "h", for "persons" or "households" respectively}
}
\value{
data.table
}
\description{
Helper to the \code{\link{pums_ftp_gofer}} function
This is used on both the person and household table to prep them for join
}
\author{
Michael Jensen
}
