% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{fetch_zip}
\alias{fetch_zip}
\title{Fetch FTP}
\usage{
fetch_zip(zip_filepath, target_files)
}
\value{
filtered input data.table with values "b" recoded to NA
}
\description{
Helper to the \code{\link{get_psrc_pums}} function
}
