% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{filter2region}
\alias{filter2region}
\title{Filter PUMS to Region}
\usage{
filter2region(dt, dyear)
}
\arguments{
\item{dt}{data.table}

\item{dyear}{The data year}
}
\value{
filtered data.table
}
\description{
Helper to \code{\link{read_pums}} function
}
\author{
Michael Jensen
}
