% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{format_acs}
\alias{format_acs}
\title{Format ACS API return}
\usage{
format_acs(dt, geography_label)
}
\arguments{
\item{dt}{data.table with Census API result}

\item{geography_label}{one of:
\itemize{
\item method1 - "block group"
\item method2 - "Tract"
\item method3 - "Place"
\item method4 - "County"
\item method5 - "MSA"
\item method6 - "public use microdata area (PUMA)"
\item method7 - "State"
}}
}
\value{
dt with fields formatted
}
\description{
Helper function for \code{\link{get_acs_single}}
}
