% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus-helpers.R
\name{format_cb_summary_tbl}
\alias{format_cb_summary_tbl}
\title{Format Census API return}
\usage{
format_cb_summary_tbl(dt, geography, census_type)
}
\arguments{
\item{dt}{data.table with Census API result}

\item{geography}{character string, i.e. "block group", "tract", "place", "county", "msa", "puma" or "state"}

\item{census_type}{character string, i.e. "acs" or "decennial"}
}
\value{
dt with fields formatted
}
\description{
Helper function for \code{\link{get_acs_single}} and \code{\link{get_decennial_single}}
}
