% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_county}
\alias{get_acs_county}
\title{ACS Estimates by County}
\usage{
get_acs_county(
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years,
  acs.type
)
}
\arguments{
\item{state}{A character string state name or abbreviation. Defaults to Washington.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}

\item{acs.type}{A character string as either 'acs1', 'acs3' or acs5'.}
}
\value{
a tibble of acs estimates by counties for selected table codes and years with a regional aggregation.
Includes detailed variable names.
}
\description{
Generate ACS estimates for multiple tables by multiple counties
}
\author{
Craig Helmann
}
