% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_msa}
\alias{get_acs_msa}
\title{ACS Estimates by MSA}
\usage{
get_acs_msa(table.names, years, acs.type, FIPS = c("14740", "42660"))
}
\arguments{
\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}

\item{acs.type}{A character string as either 'acs1', 'acs3' or acs5'.}

\item{FIPS}{Character string for FIPS codes for specific MSA geographies. Defaults to Seattle & Bremerton MSA c("14740","42660")}
}
\value{
a tibble of acs estimates by MSA for selected table codes and years. Includes detailed variable names.
}
\description{
Generate ACS estimates for multiple tables by multiple MSA's
}
\examples{
get_acs_msa(table.names=c('B03002'),
            years=c(2019),
            acs.type = 'acs1')

get_acs_msa(table.names=c('B03002'),
            years=c(2015,2017,2019),
            acs.type = 'acs1',
            FIPS = c("42660"))

get_acs_msa(table.names=c('B03002',"C17002"),
            years=c(2019),
            acs.type = 'acs1')

get_acs_msa(table.names=c('B03002',"C17002"),
            years=c(2015,2019),
            acs.type = 'acs1',
            FIPS = c("14740"))
}
\author{
Craig Helmann
}
