% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_place}
\alias{get_acs_place}
\title{ACS Estimates by Place}
\usage{
get_acs_place(
  state = "Washington",
  table.names,
  years,
  acs.type,
  place_FIPS = NULL
)
}
\arguments{
\item{state}{A character string state name or abbreviation. Defaults to Washington.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}

\item{acs.type}{A character string as either 'acs1', 'acs3' or acs5'.}

\item{place_FIPS}{Character string of FIPS codes (with state prefix) for specific Census Places. If NULL, Places within the PSRC Region will be returned.}
}
\value{
a tibble of acs estimates by Place for selected table codes and years. Includes detailed variable names.
}
\description{
Generate ACS estimates for multiple tables by multiple places
}
\author{
Craig Helmann
}
\keyword{internal}
