% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_puma}
\alias{get_acs_puma}
\title{ACS Estimates by Census PUMA}
\usage{
get_acs_puma(state = "Washington", table.names, years)
}
\arguments{
\item{state}{A character string state name or abbreviation. Defaults to Washington.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}
}
\value{
a tibble of acs estimates by PUMA for selected table codes. Includes detailed variable names.
}
\description{
Generate ACS 5 year estimates for multiple tables by Census PUMA
}
\author{
Christy Lam
}
\keyword{internal}
