% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_single}
\alias{get_acs_single}
\title{Get single ACS table}
\usage{
get_acs_single(params)
}
\arguments{
\item{params}{list of parameter combinations from \code{\link{create_param_grid}}}
}
\value{
dt formatted ACS data for a single geography and survey
}
\description{
Core function to get ACS data for any geography
}
