% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_state}
\alias{get_acs_state}
\title{ACS Estimates by State}
\usage{
get_acs_state(states = "Washington", table.names, years, acs.type)
}
\arguments{
\item{states}{A character string of state names or abbreviations. Defaults to Washington.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}

\item{acs.type}{A character string as either 'acs1', 'acs3' or acs5'.}
}
\value{
a tibble of acs estimates by states for selected table codes.
Includes detailed variable names.
}
\description{
Generate ACS estimates for multiple tables by state
This is a helper function for the main function get_acs_recs.
}
\author{
Craig Helmann
}
