% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_tract}
\alias{get_acs_tract}
\title{ACS Estimates by Census Tract}
\usage{
get_acs_tract(
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years
)
}
\arguments{
\item{state}{A character string state name or abbreviation. Defaults to Washington.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}
}
\value{
a tibble of acs estimates by tracts for selected table codes. Includes detailed variable names.
}
\description{
Generate ACS 5yr estimates for multiple tables by Census Tracts in multiple counties
}
\author{
Craig Helmann
}
