% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_msa}
\alias{get_decennial_msa}
\title{Decennial Estimates by MSA}
\usage{
get_decennial_msa(table_codes, year, fips = NULL)
}
\arguments{
\item{table_codes}{A character string or vector of Census table codes.}

\item{year}{Numeric value. A decennial year equal or greater than 2010}

\item{fips}{Character value. Single code or vector of MSA fips codes.}
}
\value{
a tibble of decennial estimates by MSA(s) for selected table codes. Does not include
variable names.
}
\description{
Generate decennial estimates for multiple tables by MSA(s).
}
\author{
Christy Lam
}
\keyword{internal}
