% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_place}
\alias{get_decennial_place}
\title{Decennial Estimates by Place}
\usage{
get_decennial_place(table_codes, years, fips = NULL, state = "WA")
}
\arguments{
\item{table_codes}{A character string or vector of Census table codes.}

\item{years}{Numeric or a vector of numeric years. A decennial year or years equal or greater than 2000.}

\item{fips}{Character value. Single code or vector of place fips codes.}

\item{state}{A character string state abbreviation}
}
\value{
a tibble of decennial estimates by place(s) for selected table codes. Does not include
variable names.
}
\description{
Generate decennial estimates for multiple tables by place(s).
}
\author{
Christy Lam
}
\keyword{internal}
