% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_recs}
\alias{get_decennial_recs}
\title{Decennial Census Estimates}
\usage{
get_decennial_recs(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  sumfile = "sf1",
  years,
  variables = NULL,
  table_codes = NULL,
  fips = NULL,
  state = "WA"
)
}
\arguments{
\item{geography}{A character string as either 'tract', 'county', 'block group', 'msa', or 'place'.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{sumfile}{A character string for which summary file to use such as "sf1" or "dp"}

\item{years}{Numeric or a vector of numeric years. A decennial year or years equal or greater than 2000.}

\item{variables}{A character string or vector of Census variables}

\item{table_codes}{A character string or vector of Census table codes,
the table code will be padded with 0s such as "H001", as opposed to "H1"}

\item{fips}{Character. For `geography = 'msa'`, provide 5-digit CBSA codes (e.g. 42660).
For `geography = 'place'`, provide state-prefixed place GEOIDs (typically 7 digits, e.g. 5363000).}

\item{state}{A character string state abbreviation. Defaults to 'WA'.}
}
\value{
a tibble of decennial estimates by selected geography for selected table codes. Includes variable names.
}
\description{
Generate decennial estimates for multiple tables by tracts, counties, MSAs, or places.
}
\examples{
tbl_names <- paste0('PCT020', LETTERS[1:6])
get_decennial_recs(geography = 'county', table_codes = tbl_names, years = 2010)

get_decennial_recs(geography = 'county', table_codes = 'P001', years = c(2000, 2010))

get_decennial_recs(geography = 'tract', table_codes = tbl_names, years = 2010)

get_decennial_recs(geography = 'place',
                   table_codes = 'PCT013',
                   years = 2010,
                   fips = c("5363000", "5308850"))

get_decennial_recs(geography = 'msa',
                   table_codes = c("H001"),
                   years = 2010,
                   fips = c('42660', "28420"))

get_decennial_recs(geography = 'block group',
                   table_codes = c('H001', 'H006'),
                   years = 2010)

get_decennial_recs(geography="tract",
                   variables="DP1_0092C",
                   years=2020, sumfile="dp")
}
\author{
Christy Lam
}
