% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_single}
\alias{get_decennial_single}
\title{Get single Decennial Census table}
\usage{
get_decennial_single(params)
}
\arguments{
\item{params}{list of parameter combinations from \code{\link{create_param_grid}}}
}
\value{
dt formatted Decennial Census data for a single geography and year
}
\description{
Core function to get Decennial Census data for any geography
}
