% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_tract_county}
\alias{get_decennial_tract_county}
\title{Decennial Estimates by Tract and County}
\usage{
get_decennial_tract_county(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table_codes,
  year,
  state = "WA"
)
}
\arguments{
\item{geography}{A character string as either 'tract', 'county', 'msa', or 'place'.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{table_codes}{A character string or vector of Census table codes.}

\item{year}{Numeric value. A decennial year equal or greater than 2010}

\item{state}{A character string state abbreviation}
}
\value{
a tibble of decennial estimates by either tracts in a county/counties for selected table codes. Does not include
variable names.
}
\description{
Generate decennial estimates for multiple tables by multiple tracts and/or counties
}
\author{
Christy Lam
}
