% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_psrc_places}
\alias{get_psrc_places}
\title{Get PSRC Places}
\usage{
get_psrc_places(years)
}
\arguments{
\item{years}{A numeric value or vector of years}
}
\value{
a data frame with relevant Census Place names and codes
}
\description{
Lookup table of Census Places within the PSRC Region, specific to the Census year
}
\author{
Mike Jensen
}
\keyword{internal}
