% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{get_psrc_pums}
\alias{get_psrc_pums}
\title{Assemble the requested PUMS data}
\usage{
get_psrc_pums(span, dyear, target_var, group_var = NULL, bin_defs = NULL)
}
\value{
A srvyr object, absent grouping but otherwise ready for summation.
}
\description{
The main PUMS assembly function.
Combines data from \code{\link{psrc_pums_targetvar}} and \code{\link{psrc_pums_groupvar}} to feed specific summary stat calls.
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
get_psrc_pums("region", 1, 2019, "AGEP", "SEX")
}
\author{
Michael Jensen
}
