% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-time-series.R
\name{get_time_series}
\alias{get_time_series}
\title{Time Series Visualization}
\usage{
get_time_series(table, variable, ts.title = NULL, print.table = FALSE)
}
\arguments{
\item{table}{Data frame of census or ACS data by Counties or Region.}

\item{variable}{Character string as Census code for a variable to plot.}

\item{ts.title}{Character. Chart title. If not specified, the title will be based on a variable name given by Census.}

\item{print.table}{Boolean. If TRUE table will be returned.}
}
\value{
a time series chart or table.
}
\description{
Generate Time Series Charts for Census or ACS Data
}
\examples{
library(rlang)
library(dplyr)
library(rlang)
library(scales)
library(stringr)
library(psrccensus)

\dontrun{
Sys.getenv("CENSUS_API_KEY")
}

ts.table = get_acs_recs(geography = 'county', table.names = c('B01001'), years=c(2019,2018,2017),
                    acs.type = 'acs5')

get_time_series(ts.table, "B01001_002")

}
\author{
Polina Butrina
}
