% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{identify_censusgeo}
\alias{identify_censusgeo}
\title{Helper to return census geography from psrccensus table}
\usage{
identify_censusgeo(df)
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}
}
\value{
string with census geography & year
}
\description{
Helper to return census geography from psrccensus table
}
\author{
Michael Jensen
}
