% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{label_decennial_variables}
\alias{label_decennial_variables}
\title{Label Decennial variables}
\usage{
label_decennial_variables(dt, year, sumfile)
}
\arguments{
\item{dt}{data.table with Census API result}

\item{year}{the year of the Decennial Census survey}

\item{sumfile}{the summary file type (e.g., 'sf1', 'dp')}
}
\value{
dataframe with labels appended
}
\description{
Helper function for \code{\link{get_decennial_single}} to provide variable labels and concept alongside codes
}
