% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bin_age}
\alias{psrc_bin_age}
\title{PSRC standard age groupings}
\usage{
psrc_bin_age(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with an additional binned age field, "BIN_AGE"
}
\description{
PSRC standard age groupings
}
