% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bin_povratio}
\alias{psrc_bin_povratio}
\title{PSRC income to poverty ratio groupings}
\usage{
psrc_bin_povratio(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with an additional binned age field, "BIN_POVRATIO"
}
\description{
PSRC income to poverty ratio groupings
}
\author{
Michael Jensen
}
