% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bin_ybl}
\alias{psrc_bin_ybl}
\title{ACS year structure built groupings}
\usage{
psrc_bin_ybl(dt, dyear)
}
\arguments{
\item{dt}{the data.table}

\item{dyear}{the data year}
}
\value{
the data.table with an additional binned age field, "BIN_YBL"
}
\description{
ACS year structure built groupings
}
