% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bincome}
\alias{psrc_bincome}
\title{PSRC standard broad income groupings}
\usage{
psrc_bincome(so)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}, with field HINCP (required)}
}
\value{
The srvyr object with an additional binned household income field, "BINCOME"
}
\description{
PSRC standard broad income groupings
}
