% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_ed_attain}
\alias{psrc_ed_attain}
\title{PSRC educational attainment groupings}
\usage{
psrc_ed_attain(dt, dyear)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with educational attainment field, "ED_ATTAIN"
}
\description{
PSRC educational attainment groupings
}
