% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{psrc_geo_convert}
\alias{psrc_geo_convert}
\title{Function to flexibly convert psrccensus estimates to any regional sf geography}
\usage{
psrc_geo_convert(df, desired_geography, desired_rpt_field, wgt = "total_pop")
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}

\item{desired_geography}{custom sf file with intended geography/geometry}

\item{desired_rpt_field}{grouping field from custom geography file, i.e. geography label}

\item{wgt}{measure share used as split weight
either "total_pop" (default), "household_pop", "group_quarters_pop", "housing_units" or "occupied_housing_units"}
}
\value{
table with custom geographic units in place of census geography units
}
\description{
Function to flexibly convert psrccensus estimates to any regional sf geography
}
\author{
Michael Jensen
}
