% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_lum_jobsector}
\alias{psrc_lum_jobsector}
\title{PSRC Land Use Modeling employment sector}
\usage{
psrc_lum_jobsector(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with modeling employment sector, "LUM_JOBSECTOR"                         # When NAICS changes, new dyear/definition set should be added
}
\description{
PSRC Land Use Modeling employment sector
}
\author{
Michael Jensen
}
