% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_lum_occupation}
\alias{psrc_lum_occupation}
\title{PSRC Modeling Occupation variable}
\usage{
psrc_lum_occupation(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with custom PSRC variable "LUM_OCCUPATION"
}
\description{
PSRC Modeling Occupation variable
}
\author{
Michael Jensen
}
