% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_mi_jobsector}
\alias{psrc_mi_jobsector}
\title{PSRC manufacturing-industrial category}
\usage{
psrc_mi_jobsector(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with manufacturing-industrial category field, "MI_JOBSECTOR"             # When NAICS changes, new dyear/definition set should be added
}
\description{
PSRC manufacturing-industrial category
}
