% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_own_rent}
\alias{psrc_own_rent}
\title{PSRC tenure groupings}
\usage{
psrc_own_rent(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with dichotomous tenure field, "OWN_RENT"
}
\description{
PSRC tenure groupings
}
