% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_ratio}
\alias{psrc_pums_ratio}
\title{Regional PUMS ratio}
\usage{
psrc_pums_ratio(span, dyear, numerator, denominator)

psrc_pums_ratio(span, dyear, numerator, denominator)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{numerator}{The exact PUMS target variable intended as numerator, as a string in UPPERCASE}

\item{denominator}{The exact PUMS variable intended as denominator, as a string in UPPERCASE}
}
\value{
A table with the variable names, regional mean, and margin of error

A table with the variable names, county ratios, and margin of error
}
\description{
This function uses psrc_pums_stat specific to the regional mean.

This function uses psrc_pums_stat specific to county ratios.
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}

\dontrun{
Sys.getenv("CENSUS_API_KEY")}

psrc_pums_ratio(1, 2019, "RNTP", "HINCP")
}
\author{
Michael Jensen

Michael Jensen
}
