% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_targetvar}
\alias{psrc_pums_targetvar}
\title{Process the PUMS Target Variable}
\usage{
psrc_pums_targetvar(span, dyear, target_var, tbl_ref)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{tbl_ref}{Either "person" or "housing", as determined by the \code{\link{get_psrc_pums}} function}
}
\value{
the filtered data.table
}
\description{
One of two primary helpers to the \code{\link{get_psrc_pums}} data assembly function.
Calls the Census API to process the primary variable used for summarization.
}
\author{
Michael Jensen
}
