% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_socp2}
\alias{psrc_socp2}
\title{2-digit SOC variable}
\usage{
psrc_socp2(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with 2-digit SOC variable, "SOCP2"
}
\description{
2-digit SOC variable
}
\author{
Michael Jensen
}
