% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_socp3}
\alias{psrc_socp3}
\title{3-digit SOC variable}
\usage{
psrc_socp3(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with 3-digit SOC variable, "SOCP3"
}
\description{
3-digit SOC variable
}
\author{
Michael Jensen
}
