% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_socp5}
\alias{psrc_socp5}
\title{5-digit SOC variable}
\usage{
psrc_socp5(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with 5-digit SOC variable, "SOCP5"
}
\description{
5-digit SOC variable
}
\author{
Michael Jensen
}
