% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_standard_jobsector}
\alias{psrc_standard_jobsector}
\title{PSRC Standard Published products employment sector}
\usage{
psrc_standard_jobsector(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with standard published covered employment sector, "STANDARD_JOBSECTOR"
}
\description{
PSRC Standard Published products employment sector
}
\author{
Michael Jensen
}
