% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus-helpers.R
\name{pull_varlist}
\alias{pull_varlist}
\title{Helper function to \code{\link{acs_varsearch}} and \code{\link{decennial_varsearch}}}
\usage{
pull_varlist(survey)
}
\arguments{
\item{survey}{tidycensus code for target survey}
}
\value{
data.table of filtered variable codes and attributes
}
\description{
Helper function to \code{\link{acs_varsearch}} and \code{\link{decennial_varsearch}}
}
