% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_bulk_stat}
\alias{pums_bulk_stat}
\title{PUMS bulk summary statistics}
\usage{
pums_bulk_stat(
  df,
  stat_type,
  target_var = NULL,
  group_var_list,
  incl_counties = TRUE
)
}
\arguments{
\item{stat_type}{Desired survey statistic}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var_list}{List with each list item a grouping variable or set of grouping variables}

\item{incl_counties}{Default=TRUE; set to false for only regional stats, e.g. if lack of data causes srvyr to give error "missing value where TRUE/FALSE needed."}

\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Generate a statistic separately for a list of grouping variables
List input items can be multiple, i.e. character vector
}
