% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_ftp_gofer}
\alias{pums_ftp_gofer}
\title{PUMS ftp go-fer}
\usage{
pums_ftp_gofer(span, dyear, level, vars, dir = NULL)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{level}{Either "p" or "h", for "persons" or "households" respectively}

\item{vars}{PUMS variable/s as an UPPERCASE string element or list}

\item{dir}{Directory for .gz file, if already downloaded. Default NULL uses the Census ftp.}
}
\value{
data.table with all requested variables,
sample & replication weights, and if needed, inflation adjustments
}
\description{
Download & filter data from the Census Bureau Microdata ftp server
Helper to the \code{\link{get_psrc_pums}} function
}
\author{
Michael Jensen
}
