% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_recode_na}
\alias{pums_recode_na}
\title{NA recode for PUMS}
\usage{
pums_recode_na(dt)
}
\arguments{
\item{dt}{data.table with Census Bureau "N/A" code ("b" or "bbb...")'}
}
\value{
filtered input data.table with values "b" recoded to NA
}
\description{
Helper to the \code{\link{get_psrc_pums}} function
}
