% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_stat}
\alias{pums_stat}
\alias{psrc_pums_count}
\alias{psrc_pums_sum}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\alias{psrc_pums_summary}
\title{PUMS summary statistics}
\usage{
psrc_pums_count(
  so,
  stat_var = NULL,
  group_vars = NULL,
  incl_na = TRUE,
  rr = FALSE
)

psrc_pums_sum(so, stat_var, group_vars = NULL, incl_na = TRUE, rr = FALSE)

psrc_pums_median(so, stat_var, group_vars = NULL, incl_na = TRUE, rr = FALSE)

psrc_pums_mean(so, stat_var, group_vars = NULL, incl_na = TRUE, rr = FALSE)

psrc_pums_summary(so, stat_var, group_vars = NULL, incl_na = TRUE, rr = FALSE)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{stat_var}{The numeric variable to summarize}

\item{group_vars}{Factor variable/s for grouping, as an UPPERCASE string element or list}

\item{incl_na}{option to remove NA from group_vars (if FALSE, the total may not reflect the full dataset)}

\item{rr}{optional relative reliability column, i.e. coefficient of variation as category levels (breakpoints: .15/.3./.5 -> good/fair/weak/unreliable)}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Separate function for total, count, median, mean
}
\examples{
\dontrun{
library(magrittr)
so <- get_psrc_pums(1, 2019, "h", c("HINCP", "TEN"))
rs <- psrc_pums_median(so, "HINCP", "TEN")
}
}
