% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_total}
\alias{psrc_pums_total}
\alias{psrc_pums_count}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\alias{county_pums_total}
\alias{county_pums_count}
\alias{county_pums_median}
\alias{county_pums_mean}
\title{PUMS summary statistics}
\usage{
psrc_pums_total(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_count(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_median(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_mean(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_total(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_count(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_median(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_mean(span, dyear, target_var, group_var = NULL, bin_defs = NULL)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}
}
\value{
A table with the variable names, summary statistic and margin of error
}
\description{
Called specific to statistic - total/count/median/mean - and to scale - region/counties
}
\section{Functions}{
\itemize{
\item \code{psrc_pums_total}: Generate regional PUMS totals

\item \code{psrc_pums_count}: Generate regional PUMS count

\item \code{psrc_pums_median}: Generate regional PUMS median

\item \code{psrc_pums_mean}: Generate regional PUMS mean

\item \code{county_pums_total}: Generate PUMS totals by county

\item \code{county_pums_count}: Generate PUMS counts <tally> by county

\item \code{county_pums_median}: Generate PUMS medians by county

\item \code{county_pums_mean}: Generate PUMS averages <mean> by county
}}

\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
psrc_pums_total(1, 2019, "AGEP", "SEX")
county_pums_median(1, 2019, "AGEP", "SEX")

}
\author{
Michael Jensen
}
