% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_varsearch}
\alias{pums_varsearch}
\title{Search PUMS variable definitions}
\usage{
pums_varsearch(regex)
}
\value{
data.table of filtered variable attributes
}
\description{
Look for a var_code via search term
}
