% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{real_dollars}
\alias{real_dollars}
\title{Inflation adjustment to a year other than the survey year}
\usage{
real_dollars(so, refyear)
}
\arguments{
\item{so}{srvyr object returned by get_psrc_pums}

\item{refyear}{dollar year in which value should be returned be returned}
}
\description{
Applies PCE deflator to yield real values in specified dollar-year terms
Helpful when comparing between separate surveys
Requires [St. Louis Federal Reserve (FRED) API key](http://sboysel.github.io/fredr/articles/fredr.html#authentication)
}
\author{
Michael Jensen
}
