% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{recode_dt}
\alias{recode_dt}
\title{Recode factor variables to labels}
\usage{
recode_dt(rcd)
}
\arguments{
\item{rcd}{variable to be recoded}
}
\value{
data.table with factor label instead of value
}
\description{
Helper to \code{\link{get_psrc_pums}} function.
Swaps factor values for labels.
}
