% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{regional_pums_stat}
\alias{regional_pums_stat}
\alias{psrc_pums_total}
\alias{psrc_pums_count}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\title{Regional PUMS summary statistics}
\usage{
psrc_pums_total(so, target_var, group_vars = NULL)

psrc_pums_count(so, group_vars = NULL)

psrc_pums_median(so, target_var, group_vars = NULL)

psrc_pums_mean(so, target_var, group_vars = NULL)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{Factor variable/s for grouping, as an UPPERCASE string element or list}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Separate function for total, count, median, or mean
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
get_psrc_pums(1, 2019, "h", "HINCP", "TEN") \%>\% psrc_pums_median("HINCP", "TEN")

}
