% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{reliability_calcs}
\alias{reliability_calcs}
\title{Add estimate reliability information}
\usage{
reliability_calcs(dfs, moe = "moe", estimate = "estimate")
}
\arguments{
\item{dfs}{data frames already retrieved from the api, or created by a user}

\item{moe}{character name of the column that has contains margin of error estimates}

\item{estimate}{character name of the column that has contains data value estimates}
}
\value{
the data frames with new columns for se, cv, and reliability
}
\description{
After gathering data, add reliability information using moe and estimate
}
\author{
Suzanne
}
