% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{test_df_for_conversion}
\alias{test_df_for_conversion}
\title{Helper to warn if input table doesn't meet criteria for geographic conversion}
\usage{
test_df_for_conversion(df)
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}
}
\value{
NULL (if fails); 1 (if succeeds)
}
\description{
Helper to warn if input table doesn't meet criteria for geographic conversion
}
\author{
Michael Jensen
}
