% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{testInteger}
\alias{testInteger}
\title{Helper to ascertain datatype of variable}
\usage{
testInteger(x)
}
\arguments{
\item{x}{vector, e.g. variable}
}
\value{
string with census geography & year
}
\description{
Helper to ascertain datatype of variable
}
