% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{verify_custom_geo}
\alias{verify_custom_geo}
\title{Helper to verify custom geometry input meets criteria for geographic conversion}
\usage{
verify_custom_geo(custom_geo, custom_geo_var)
}
\arguments{
\item{custom_geo}{custom sf file with intended geography/geometry}

\item{custom_geo_var}{grouping field from custom geography file, i.e. geography label}
}
\value{
verified sf object
}
\description{
Helper to verify custom geometry input meets criteria for geographic conversion
}
\author{
Michael Jensen
}
