% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{verify_custom_geography}
\alias{verify_custom_geography}
\title{Helper to verify custom geometry input meets criteria for geographic conversion}
\usage{
verify_custom_geography(desired_geography, desired_rpt_field)
}
\arguments{
\item{desired_geography}{custom sf file with intended geography/geometry}

\item{desired_rpt_field}{grouping field from custom geography file, i.e. geography label}
}
\value{
verified sf object
}
\description{
Helper to verify custom geometry input meets criteria for geographic conversion
}
\author{
Michael Jensen
}
