% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{write_pums_rds}
\alias{write_pums_rds}
\title{Save PUMS data for offline access}
\usage{
write_pums_rds(dyear, dir)
}
\arguments{
\item{dyear}{The data year}

\item{dir}{Directory for .gz file}
}
\value{
Message; saves separate .rds files in specified location
}
\description{
Save PUMS data for offline access
}
\author{
Michael Jensen
}
