% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{z_score}
\alias{z_score}
\title{Z Score}
\usage{
z_score(x, y)
}
\arguments{
\item{x}{numeric vector, first estimate and corresponding MOE to compare}

\item{y}{numeric vector, second estimate and corresponding MOE to compare}
}
\value{
Z score; if larger than 1, difference is significant
}
\description{
Stat to determine if two estimates are different
}
\author{
Michael Jensen
}
