% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{acs_varsearch}
\alias{acs_varsearch}
\title{Search published ACS variables}
\usage{
acs_varsearch(regex, year = NULL)
}
\arguments{
\item{regex}{search term}

\item{year}{optionally restrict search to a specific year}
}
\value{
data.table of filtered variable codes and attributes
}
\description{
Identify desired tables by examining prefix of relevant variable codes
Includes primary, subject and profile tables
}
\author{
Michael Jensen
}
