% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{add_county}
\alias{add_county}
\title{Add County Name to PUMS API Result}
\usage{
add_county(dt)
}
\arguments{
\item{dt}{The data table}
}
\value{
the filtered data.table
}
\description{
This helper function is used in both the psrc_pums_groupvar and psrc_pums_targetvar functions.
It attaches the county name.
}
\author{
Michael Jensen
}
