% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{add_decennial_share}
\alias{add_decennial_share}
\title{Add shares to Decennial Census result}
\usage{
add_decennial_share(df)
}
\arguments{
\item{df}{dataframe with Decennial Census result}
}
\value{
dataframe with additional share field
}
\description{
Add shares to Decennial Census result
}
\author{
Michael Jensen
}
