% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{adjust_dollars}
\alias{adjust_dollars}
\title{Dollar variable adjustment for PUMS}
\usage{
adjust_dollars(dt, dollar_var)
}
\arguments{
\item{dt}{The data table}

\item{dollar_var}{The reported PUMS variable to adjust}
}
\value{
the data.table with dollar value adjusted
}
\description{
This helper function is used in both the psrc_pums_groupvar and psrc_pums_targetvar functions.
It matches the corresponding adjustment factor with the appropriate PUMS variable and applies the adjust_inflation function.
}
\author{
Michael Jensen
}
