% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{adjust_dollars}
\alias{adjust_dollars}
\title{Dollar variable adjustment for PUMS}
\usage{
adjust_dollars(dt, dollar_var)
}
\arguments{
\item{dt}{The data.table}

\item{dollar_var}{PUMS variable to adjust}
}
\value{
full input data.table with all dollar values adjusted
}
\description{
Helper to \code{\link{psrc_pums_groupvar}} and \code{\link{psrc_pums_targetvar}} functions.
Matches the corresponding adjustment factor with the appropriate PUMS variable and applies the \code{\link{adjust_inflation}} function.
}
\author{
Michael Jensen
}
