% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{adjust_inflation}
\alias{adjust_inflation}
\title{Inflation adjustment for PUMS}
\usage{
adjust_inflation(dt, dollar_var, adj_var)
}
\arguments{
\item{dt}{The data table as conditioned by adjust_dollars function}

\item{dollar_var}{The reported PUMS variable to adjust}

\item{adj_var}{The PUMS adjustment factor: either ADJINC for income or ADJHSG for housing cost}
}
\value{
the data.table with the input variable adjusted
}
\description{
This is a helper function for the adjust_dollars function
}
\author{
Michael Jensen
}
