% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{census_to_psrcgeo}
\alias{census_to_psrcgeo}
\alias{census_to_rgs}
\alias{census_to_rgc}
\alias{census_to_mic}
\alias{census_to_taz}
\title{Translate psrccensus data to planning geographies}
\usage{
census_to_rgs(df, wgt = "total_pop")

census_to_rgc(df, wgt = "total_pop")

census_to_mic(df, wgt = "total_pop")

census_to_taz(df, wgt = "total_pop")
}
\arguments{
\item{df}{acs or dicennial dataset returned from psrccensus}

\item{wgt}{either "total_pop" (default), "household_pop", "group_quarters_pop", "housing_units" or "occupied_housing_units"}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Translate psrccensus data to planning geographies

Translate psrccensus data to Regional Geography Class (RGS)

Translate psrccensus data to Regional Growth Center (RGC)

Translate psrccensus data to Regional Manufacturing-Industrial Center (MIC)

Translate psrccensus data to Traffic Analysis Zone (TAZ)
}
