% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{clip2region}
\alias{clip2region}
\title{Restrict PUMS API pull to PSRC region}
\usage{
clip2region(dt)
}
\arguments{
\item{dt}{Washington State PUMS data.table}
}
\value{
PSRC data.table
}
\description{
Helper to \code{\link{psrc_pums_groupvar}} and \code{\link{psrc_pums_targetvar}} functions.
Filters the Washington State API pull to the PSRC region and removes State/State_label columns.
}
\author{
Michael Jensen
}
