% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{county_pums_stat}
\alias{county_pums_stat}
\alias{county_pums_total}
\alias{county_pums_count}
\alias{county_pums_median}
\alias{county_pums_mean}
\title{PUMS summary statistics by county}
\usage{
county_pums_total(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_count(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_median(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

county_pums_mean(span, dyear, target_var, group_var = NULL, bin_defs = NULL)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}
}
\value{
A table with the counties, variable names and labels, summary statistic, and margin of error
}
\description{
Separate function for total, count, median, or mean
}
\examples{
\dontrun{Sys.getenv("CENSUS_API_KEY")}
county_pums_median(span=1, dyear=2019, target_var="AGEP", group_var="SEX")

}
\author{
Michael Jensen
}
