% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{create_param_grid}
\alias{create_param_grid}
\title{Generate parameter list for all combinations}
\usage{
create_param_grid(table.names, years, params)
}
\arguments{
\item{table.names}{Vector of table names}

\item{years}{Vector of years}

\item{params}{Base parameters to include with each combination}
}
\value{
List of parameter combinations
}
\description{
Helper function for \code{\link{get_acs_recs}}; creates a list of parameter combinations for all tables and years
}
