% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tract-map.R
\name{create_tract_map}
\alias{create_tract_map}
\title{Census Tract Mapping}
\usage{
create_tract_map(
  tract.tbl,
  tract.lyr,
  map.title = NULL,
  map.subtitle = NULL,
  map.title.position = NULL,
  legend.title = NULL,
  legend.subtitle = NULL,
  map.lat = 47.615,
  map.lon = -122.257,
  map.zoom = 8.5,
  wgs84 = 4326
)
}
\arguments{
\item{tract.tbl}{A data frame of census data by Census Tract to map.}

\item{tract.lyr}{A spatial layer of Census Tracts.}

\item{map.title}{Character. Map title.}

\item{map.subtitle}{Character. A map subtitle that will appear below the map title.}

\item{map.title.position}{Character. Place the map title and subtitle in 'bottomleft','bottomright', topleft', or 'topright'.}

\item{legend.title}{Character. Legend title.}

\item{legend.subtitle}{Character. A legend subtitle that will appear below the legend subtitle.}

\item{map.lat}{A numeric value for the latitude of the center point for your map. Defaults to 47.615 (PSRC Region)}

\item{map.lon}{A numeric value for the longitude of the center point for your map. Defaults to -122.257 (PSRC Region)}

\item{map.zoom}{A numeric value for the default zoom level for your map. Defaults to 8.5 (PSRC Region)}

\item{wgs84}{A code for mapping, this should be pulled out of the function or something}
}
\value{
a leaflet map object.
}
\description{
Generate Census Tract Maps for Census Data
}
\examples{
library(sf)
library(dplyr)
\dontrun{
Sys.getenv("CENSUS_API_KEY")
}

tract.big.tbl <- psrccensus::get_acs_recs(geography='tract',table.names=c('B03002'),years=c(2019))
tract.tbl <- tract.big.tbl \%>\%
filter(label=='Estimate!!Total:!!Not Hispanic or Latino:!!Black or African American alone')

gdb.nm <- paste0("MSSQL:server=",
"AWS-PROD-SQL\\\\Sockeye",
";database=",
"ElmerGeo",
";trusted_connection=yes")

spn <-  2285

tract_layer_name <- "dbo.tract2010_nowater"

tract.lyr <- st_read(gdb.nm, tract_layer_name, crs = spn)

create_tract_map(tract.tbl=tract.tbl, tract.lyr=tract.lyr,
map.title='Black, non-Hispanic Population',
,map.title.position='topleft', legend.title='Black, Non-Hispanic Population',
legend.subtitle='by Census Tract')
}
\author{
Suzanne Childress
}
