% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{fetch_zip}
\alias{fetch_zip}
\title{Fetch ZIP}
\usage{
fetch_zip(zip_filepath, target_files, dyear)
}
\arguments{
\item{zip_filepath}{ftp URL location}

\item{target_files}{.zip archive file to read}

\item{dyear}{The data year}
}
\value{
unzipped table
}
\description{
Helper to the \code{\link{pums_ftp_gofer}} function
}
