% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{get_acs_recs}
\alias{get_acs_recs}
\title{ACS Estimates}
\usage{
get_acs_recs(
  geography,
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years,
  FIPS = c("14740", "42660"),
  place_FIPS = NULL,
  acs.type
)
}
\arguments{
\item{geography}{A character string as either 'tract', 'county', 'msa', or 'place'.}

\item{state}{A character string state name or abbreviation. Defaults to Washington.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{table.names}{A character string or vector of Census table codes.}

\item{years}{A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.}

\item{FIPS}{Character string for FIPS codes for specific MSA geographies. Defaults to Seattle & Bremerton MSA c("14740","42660")}

\item{acs.type}{A character string as either 'acs1', 'acs3' or acs5'.}
}
\value{
a tibble of ACS estimates by selected geography for selected table codes. Includes variable names.
}
\description{
Generate ACS estimates for multiple tables by tracts, counties, MSAs, or places for multiple years.
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
get_acs_recs(geography = 'county',
             table.names = c('B03002',"C17002"),
             years=c(2017,2019),
             acs.type = 'acs1')

get_acs_recs(geography = 'county',
             table.names = c('B03002'),
             counties=c("Kitsap"),
             years=c(2019),
             acs.type = 'acs5')

get_acs_recs(geography = 'msa',
             table.names = c('B03002',"C17002"),
             years=c(2019),
             acs.type = 'acs1')

get_acs_recs(geography = 'msa',
             table.names = c("C17002"),
             years=c(2019),
             FIPS = c("14740"),
             acs.type = 'acs5')

get_acs_recs(geography = 'place',
             table.names = c("C17002"),
             years=c(2019),
             acs.type = 'acs5')

get_acs_recs(geography = 'tract',
             counties=c("Pierce"),
             table.names = c('B03002'),
             years=c(2018,2019))


}
\author{
Craig Helmann
}
