% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_place}
\alias{get_decennial_place}
\title{Decennial Estimates by Place}
\usage{
get_decennial_place(table_codes, year, fips = NULL, state = "WA")
}
\arguments{
\item{table_codes}{A character string or vector of Census table codes.}

\item{year}{Numeric value. A decennial year equal or greater than 2010}

\item{fips}{Character value. Single code or vector of place fips codes.}

\item{state}{A character string state abbreviation}
}
\value{
a tibble of decennial estimates by place(s) for selected table codes. Does not include
variable names.
}
\description{
Generate decennial estimates for multiple tables by place(s).
}
\examples{
Sys.getenv("CENSUS_API_KEY")
get_decennial_place(table_codes = 'PCT013', year = 2010)

get_decennial_place(table_codes = 'PCT013',
                    year = 2010,
                    fips = c("5363000", "5308850"))

get_decennial_place(table_codes = c('PCT013', 'PCT022'),
                    year = 2010,
                    fips = c("5363000", "5308850"))
}
\author{
Christy Lam
}
