% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_recs}
\alias{get_decennial_recs}
\title{Decennial Estimates}
\usage{
get_decennial_recs(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table_codes,
  year,
  fips = NULL
)
}
\arguments{
\item{geography}{A character string as either 'tract', 'county', 'msa', or 'place'.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{table_codes}{A character string or vector of Census table codes.}

\item{year}{Numeric. A decennial year equal or greater than 2010.}

\item{fips}{Character. Single code or vector of either MSA or place fips codes.}
}
\value{
a tibble of decennial estimates by selected geography for selected table codes. Includes variable names.
}
\description{
Generate decennial estimates for multiple tables by tracts, counties, MSAs, or places.
}
\examples{
Sys.getenv("CENSUS_API_KEY")
tbl_names <- paste0('PCT020', LETTERS[1:6])
get_decennial_recs(geography = 'county', table_codes = tbl_names, year = 2010)

get_decennial_recs(geography = 'tract', table_codes = tbl_names, year = 2010)

get_decennial_recs(geography = 'place',
                   table_codes = 'PCT013',
                   year = 2010,
                   fips = c("5363000", "5308850"))

get_decennial_recs(geography = 'msa',
                   table_codes = c("H001", "P001"),
                   year = 2010,
                   fips = c('42660', "28420"))
}
\author{
Christy Lam
}
