% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-decennial-recs.R
\name{get_decennial_tract_county_bg}
\alias{get_decennial_tract_county_bg}
\title{Decennial Estimates by Tract or Block Group and County}
\usage{
get_decennial_tract_county_bg(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  variables,
  table_codes,
  years,
  state = "WA",
  sumfile
)
}
\arguments{
\item{geography}{A character string as either 'tract', 'county', 'block group'.}

\item{counties}{A character string or vector of counties. Defaults to PSRC counties.}

\item{variables}{A character string or vector of Census variables}

\item{table_codes}{A character string or vector of Census table codes.}

\item{years}{Numeric or a vector of numeric years. A decennial year or years equal or greater than 2000.}

\item{state}{A character string state abbreviation}

\item{sumfile}{A character string for which summary file to use such as "sf1" or "dp"}
}
\value{
a tibble of decennial estimates by either tracts in a county/counties for selected table codes and years. Does not include
variable names.
}
\description{
Generate decennial estimates for multiple tables by multiple tracts and/or counties
}
\author{
Christy Lam
}
\keyword{internal}
