% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-equity-geographies.R
\name{get_equity_geographies}
\alias{get_equity_geographies}
\title{Get equity tracts}
\usage{
get_equity_geographies(
  equity_type = "50",
  equity_group = "race",
  year = "2019"
)
}
\arguments{
\item{equity_type}{string, The type of equity geography either '50' (50 percent) or 'regavg' (regional average)}

\item{equity_group}{string, The group of people described by the equity geography, options are:
'race',
'poverty',
'disability',
'elderly',
'limited_english'}

\item{year}{string, the year for which the equity geography is defined}
}
\value{
a data frame with all tracts in the region with a 0-1 variable for whether they are in an equity geography or not
}
\description{
Retrieve a 0-1 variable for whether each tract in the region is in an equity geography or not
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")
}

get_equity_geographies(equity_type = 'regavg', equity_group = 'disability', year = '2019')

}
\author{
Suzanne Childress
}
