% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{get_psrc_places}
\alias{get_psrc_places}
\title{Retrieve set of Census Places within the PSRC region}
\usage{
get_psrc_places(year)
}
\arguments{
\item{year}{of census geography}
}
\value{
string vector of Place FIPS codes
}
\description{
Retrieve set of Census Places within the PSRC region
}
