% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{get_psrc_pums}
\alias{get_psrc_pums}
\title{Assemble the requested PUMS data}
\usage{
get_psrc_pums(span, dyear, target_var, group_var = NULL, bin_defs = NULL)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}
}
\value{
A srvyr object ready for summation.
}
\description{
This is the main assembly function.
It draws from psrc_pums_targetvar and psrc_pums_groupvar and feeds the specific summary stat calls.
}
\examples{
library(dplyr)
library(srvyr)
\dontrun{
Sys.getenv("CENSUS_API_KEY")}

get_psrc_pums("region", 1, 2019, "AGEP", "SEX")
}
\author{
Michael Jensen
}
