% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{get_psrc_pums}
\alias{get_psrc_pums}
\title{Retrieve and assemble PUMS data}
\usage{
get_psrc_pums(
  span,
  dyear,
  level,
  vars,
  dollar_adj = TRUE,
  dir = NULL,
  labels = TRUE
)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{level}{Either "p" or "h", for "persons" or "households" respectively}

\item{vars}{PUMS variable/s as an UPPERCASE string element or list}

\item{dollar_adj}{Default TRUE; adjust income and cost values for inflation using the averaged factors provided with PUMS}

\item{dir}{Directory for .gz file, if already downloaded. Default NULL uses the Census ftp.}

\item{labels}{Default TRUE, recodes varible values to the corresponding label}
}
\value{
A srvyr object with appropriate sampling weight and replication weights
}
\description{
The primary PUMS function
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")
get_psrc_pums(span=1, dyear=2019, level="p", vars=c("AGEP","SEX"))}

}
