% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{label_acs_variables}
\alias{label_acs_variables}
\title{Label ACS variables}
\usage{
label_acs_variables(df, table.name, year, acs.type)
}
\arguments{
\item{df}{dataframe with Census API result}

\item{table.name}{Census table code}

\item{year}{the year--or last year--of the ACS survey}

\item{acs.type}{either acs1 or acs5}
}
\value{
dataframe with labels appended
}
\description{
Helper function to provide variable labels and concept--i.e. topic--alongside codes
}
