% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bin_age}
\alias{psrc_bin_age}
\title{PSRC standard age groupings}
\usage{
psrc_bin_age(so)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}, with field AGEP (required)}
}
\value{
The srvyr object with an additional binned age field, "BINAGE"
}
\description{
PSRC standard age groupings
}
