% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-groupings.R
\name{psrc_bincome}
\alias{psrc_bincome}
\title{PSRC standard broad income groupings}
\usage{
psrc_bincome(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with an additional binned household income field, "BINCOME"
}
\description{
PSRC standard broad income groupings
}
