% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_groupvar}
\alias{psrc_pums_groupvar}
\title{Process the PUMS Grouping Variable}
\usage{
psrc_pums_groupvar(span, dyear, group_var, tbl_ref, key_ref, bin_defs)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{tbl_ref}{PUMS table (either "person" or "housing") of the target variable}

\item{key_ref}{PUMS table (either "person" or "housing") of the grouping variable}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}
}
\value{
the filtered data.table
}
\description{
One of two primary helpers to the \code{\link{get_psrc_pums}} data assembly function.
Calls the Census API to process the optional grouping variable.
}
\author{
Michael Jensen
}
