% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-pums-stat.R
\name{psrc_pums_stat}
\alias{psrc_pums_stat}
\title{Generic call for PUMS summary statistics}
\usage{
psrc_pums_stat(so, stat_type, stat_var, group_vars, incl_na = TRUE, rr = FALSE)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{stat_type}{Desired survey statistic}

\item{stat_var}{The numeric variable to summarize}

\item{group_vars}{Factor variable/s for grouping, as an UPPERCASE string element or list}

\item{incl_na}{option to remove NA from group_vars (if FALSE, the total may not reflect the full dataset)}

\item{rr}{optional relative reliability column, i.e. coefficient of variation as category levels (breakpoints: .15/.3./.5 -> good/fair/weak/unreliable)}
}
\value{
A summary tibble, including variable names, summary statistic and margin of error
}
\description{
Given specific form by related \code{\link{pums_stat}} functions.
}
\author{
Michael Jensen
}
