% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_stat}
\alias{psrc_pums_stat}
\title{Generic call for PUMS summary statistics}
\usage{
psrc_pums_stat(
  stat_type,
  geo_scale,
  span,
  dyear,
  target_var,
  group_var,
  bin_defs
)
}
\arguments{
\item{stat_type}{Desired survey statistic}

\item{geo_scale}{Either "county" or "region"}
}
\value{
A table with the variable names, desired statistic, and margin of error
}
\description{
Given specific form by related total/count/median/mean x region/county \code{\link{pumstat}} functions.
Fed from \code{\link{get_psrc_pums}}
}
\author{
Michael Jensen
}
