% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{psrc_pums_stat}
\alias{psrc_pums_stat}
\title{Generic call for PUMS summary statistics}
\usage{
psrc_pums_stat(so, stat_type, target_var, group_vars)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{stat_type}{Desired survey statistic}

\item{target_var}{Numeric PUMS analysis variable, as an UPPERCASE string}

\item{group_var}{Factor variable/s for grouping, as an UPPERCASE string element or list}
}
\value{
A summary tibble, including variable names, summary statistic and margin of error
}
\description{
Given specific form by related \code{\link{pums_stat}} functions.
}
