% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus-details.R
\docType{package}
\name{psrccensus}
\alias{psrccensus}
\alias{psrccensus-package}
\title{psrccensus: A package for easy compilation of census estimates for the Puget Sound Region.}
\description{
This package provides functions to easily collect census estimates across geographies within the
Central Puget Sound region. It works with Decennial, ACS and PUMS datasets
}
\author{
Christy Lam \email{clam@psrc.org}
}
