% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_api_gofer}
\alias{pums_api_gofer}
\title{PUMS API go-fer}
\usage{
pums_api_gofer(span, dyear, level, vars)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{level}{Either "p" or "h", for "persons" or "households" respectively}

\item{vars}{PUMS variable/s as an UPPERCASE string element or list}
}
\value{
data.table with all requested variables,
sample & replication weights, and if needed, inflation adjustments
}
\description{
Call the Census Bureau Microdata API
Helper to the \code{\link{get_psrc_pums}} function
}
\author{
Michael Jensen
}
