% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{pums_stat}
\alias{pums_stat}
\alias{psrc_pums_count}
\alias{psrc_pums_sum}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\alias{psrc_pums_summary}
\title{PUMS summary statistics}
\usage{
psrc_pums_count(so, stat_var = NULL, group_vars = NULL)

psrc_pums_sum(so, stat_var, group_vars = NULL)

psrc_pums_median(so, stat_var, group_vars = NULL)

psrc_pums_mean(so, stat_var, group_vars = NULL)

psrc_pums_summary(so, stat_var, group_vars = NULL)
}
\arguments{
\item{so}{The srvyr object returned by \code{\link{get_psrc_pums}}}

\item{stat_var}{The numeric variable to summarize}

\item{group_vars}{Factor variable/s for grouping, as an UPPERCASE string element or list}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Separate function for total, count, median, mean
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")
library(magrittr)
so <- get_psrc_pums(1, 2019, "h", c("HINCP", "TEN"))
rs <- psrc_pums_median(so, "HINCP", "TEN")
}
}
