% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{read_pums}
\alias{read_pums}
\title{Read PUMS csv}
\usage{
read_pums(target_file, dyear)
}
\arguments{
\item{dyear}{The data year}

\item{csv_target}{csv filename}
}
\value{
unzipped table
}
\description{
Helper to the \code{\link{pums_ftp_gofer}} function
Can be either .csv from ftp site, or compressed .gz file on server
}
