% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{regional_pums_stat}
\alias{regional_pums_stat}
\alias{psrc_pums_total}
\alias{psrc_pums_count}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\title{Regional PUMS summary statistics}
\usage{
psrc_pums_total(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_count(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_median(span, dyear, target_var, group_var = NULL, bin_defs = NULL)

psrc_pums_mean(span, dyear, target_var, group_var = NULL, bin_defs = NULL)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Separate function for total, count, median, or mean
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
psrc_pums_median(1, 2019, "AGEP", "SEX")

}
\author{
Michael Jensen
}
