% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-psrc-pums.R
\name{regional_pums_stat}
\alias{regional_pums_stat}
\alias{psrc_pums_total}
\alias{psrc_pums_count}
\alias{psrc_pums_median}
\alias{psrc_pums_mean}
\title{Regional PUMS summary statistics}
\usage{
psrc_pums_total(
  span,
  dyear,
  target_var,
  group_var = NULL,
  bin_defs = NULL,
  dollar_adj = TRUE
)

psrc_pums_count(
  span,
  dyear,
  target_var,
  group_var = NULL,
  bin_defs = NULL,
  dollar_adj = TRUE
)

psrc_pums_median(
  span,
  dyear,
  target_var,
  group_var = NULL,
  bin_defs = NULL,
  dollar_adj = TRUE
)

psrc_pums_mean(
  span,
  dyear,
  target_var,
  group_var = NULL,
  bin_defs = NULL,
  dollar_adj = TRUE
)
}
\arguments{
\item{span}{Either 1 for acs1 or 5 for acs5}

\item{dyear}{The data year}

\item{target_var}{The exact PUMS target variable intended, as a string in UPPERCASE}

\item{group_var}{The exact PUMS variable intended for grouping, as a string in UPPERCASE}

\item{bin_defs}{Optional argument: if a single number, used as Ntile; if a list, used as custom bin breakpoints}

\item{dollar_adj}{Default TRUE; adjust income and cost values for inflation using the averaged factors provided with PUMS}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
Separate function for total, count, median, or mean
}
\examples{
\dontrun{
Sys.getenv("CENSUS_API_KEY")}
psrc_pums_median(span=1, dyear=2019, target_var="AGEP", group_var="SEX")

}
\author{
Michael Jensen
}
