% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{reliability_calcs}
\alias{reliability_calcs}
\title{Add estimate reliability information}
\usage{
reliability_calcs(dfs)
}
\arguments{
\item{dfs}{acs data frames already retrieved from the api}
}
\value{
the data frames with new columns for se, cv, and reliability
}
\description{
After gathering data, add reliability information using moe and estimate
}
\author{
Suzanne
}
