% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-recs.R
\name{reliability_calcs}
\alias{reliability_calcs}
\title{Add estimate reliability metrics}
\usage{
reliability_calcs(df, estimate = "estimate", moe = "moe")
}
\arguments{
\item{df}{dataframe or data.table already retrieved from the api, or created by a user}

\item{estimate}{character name of the column that has contains data value estimates}

\item{moe}{character name of the column that has contains margin of error estimates}
}
\value{
the data frames with new columns for se, cv, and reliability
}
\description{
Add coefficient of variation (cv) and relative reliability rating to estimates
Census ACS uses 90 percent confidence interval, i.e. z score 1.645
http://aws-linux/mediawiki/index.php/Understanding_Error_and_Determining_Statistical_Significance
}
\author{
Suzanne Childress
}
