% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{use_geography_splits}
\alias{use_geography_splits}
\title{Helper to translate psrccensus estimates to planning geographies}
\usage{
use_geography_splits(df, planning_geog_type, wgt = "total_pop")
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}

\item{planning_geog_type}{planning geography type as listed in Elmer.general.geography_splits}

\item{wgt}{measure share used as split weight
either "total_pop" (default), "household_pop", "group_quarters_pop", "housing_units" or "occupied_housing_units"}
}
\value{
table with planning geography units in place of census geography units
}
\description{
Helper to translate psrccensus estimates to planning geographies
}
\author{
Michael Jensen
}
