% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrccensus_geographies.R
\name{use_geography_splits}
\alias{use_geography_splits}
\title{Helper to translate psrccensus estimates to planning geographies}
\usage{
use_geography_splits(
  df,
  planning_geog_type,
  wgt,
  ofm_vintage = "default",
  parcel_year = "default"
)
}
\arguments{
\item{df}{acs or decennial dataset returned from psrccensus}

\item{planning_geog_type}{planning geography type as listed in Elmer.general.geography_splits}

\item{wgt}{measure share used as split weight
either "total_pop", "household_pop", "group_quarters_pop", "housing_units" or "occupied_housing_units"}

\item{ofm_vintage}{for deprecated splits; otherwise keep default. See <http://aws-linux/mediawiki/index.php/Get_geography_splits_(Elmer_Function)>}

\item{parcel_year}{for deprecated baseyear; otherwise keep default. See <http://aws-linux/mediawiki/index.php/Get_geography_splits_(Elmer_Function)>}
}
\value{
table with planning geography units in place of census geography units
}
\description{
Helper to translate psrccensus estimates to planning geographies
}
\author{
Michael Jensen
}
